==> 3. Random_selector.ipynb <==
From the analysis above the collection of documents from the Nexis Uni was skewed. To mitigate this issue, only the 70th Percentile articles - 203 articles was used
The article is run interatively with same random seed to ensure reproduceability

import random

# Set random seed for reproducibility
random.seed(21)
threshold = 203
total_article_count = 14883
# Variables to modify

total_articles = 608 # Update this number for the articles you want to select


calculated_articles = round((total_articles/total_article_count)*threshold,0)
num_random_numbers = int(calculated_articles)

# Generate random numbers within the range of articles
# Article numbers from 1 to total_articles
random_article_numbers = random.sample(range(1, total_articles + 1), num_random_numbers)

# Sort the numbers for easier reading
random_article_numbers.sort()

print("Random Article Selection")
print("=" * 25)
print("Total articles:", total_articles)
print("Numbers requested:", num_random_numbers)
print("Random seed:", 21)
print()
print("Selected article numbers:", random_article_numbers)
