==> 6.Stopword List.ipynb <==
Stopword List

custom_stopwords = {
    # Very high frequency, low descriptive value verbs
    "say", "tell", "make", "get", "go", "take", "come", "see", "know", "think",
    "want", "use", "find", "give", "ask", "call", "try", "look", "seem", "believe",
    "start", "set", "put", "add", "move", "include", "offer", "present", "return",
    "leave", "remain", "join", "attend", "work", "write", "become", "appear",
    "respond", "decide", "hold", "raise", "pick", "replace", "buy", "describe",
    "post", "joke", "echo", "watch", "film", 
    "serve", "push", "embrace", "argue", "seek", "bring", "succeed", "donate",
    # ---------------
    "bother", "build", "oust", "recall", "spend", "decline", "reveal", "draw",
    "name", "reassure", "testify", "create", "oversee", "walk", "invest",
    "also", "still", "later", "back", "early", "often", "far", "shortly", "ago",
    "recently", "long", "old", "continue", "consider", "found", "stand", "reach", "step", "grow", "drop",
    "retain", "begin", "reply", "keep", "hope", "sell", "meet",
    "issue", "point", "read", "address", "agree", "send", "receive", "talk",
    "need", "feel", "expect", "pull", "speak", "turn", "show", "like", "involve",
    "lead", "suggest", "sit", "hand", "discuss", "promise", "bear", "thank",
    "welcome", "develop", "follow", "indicate", "run", "note", "pay", "play",
    #------------------
    "claim", "lose", "mention", "already", "reportedly", "however", "previously", "first", 
    "continue", "consider", "found", "stand", "reach", "grow", "drop", "retain",
    "begin", "reply", "keep", "hope", "sell", "meet", "issue",
    "point", "read", "address", "agree", "send", "receive", "talk", "need",
    "feel", "expect", "pull", "turn", "show", "like", "involve", "sit", "hand",
    "discuss", "promise", "bear", "thank", "welcome", "develop", "follow",
    "indicate", "run", "note", "pay", "play", "claim", "lose", "mention",
    # ------------------------
    "visit", "lay","ahead", "ultimately", "well", "publicly", "however", "even", "next",
    "effectively", "quickly", "instead", "away", "ever", "really", "immediately",
    "person", "company", "man", "son", "zuckerberg", "khosla", "figure", "public", "sundar", "arnault", "live", "answer", 
    "compare", "introduce", "refer", "hit", "sign", "hint", "enter", "end", "change", "stay", "report", 
    "well", "publicly", "eventually", "simply", "yet", "privately", "quickly", "instead", "immediately",
    "chief", "manager", "deputy", "boss", "director"
    "company", "page", "zuckerberg", "co", "urge", "marry", "remove", "november","earn","learn","satyanadella", "altman", "consultant", "easily", "eat", "brockman", "file", "cite", "easily", "target", 
    "team", "unveil","feature","eld", "count", "largely", "briefly","rank","cook","photograph",
    #---------------------------
    "face","part", "always", "directly", "chair", "stop" , "officer", "wear" ,"onstage", "opt", "indeed","meanwhile","break","interview","email","tweet","die","hail"
    "officer", "executive" ,"chairman","ceo", "president", "investor", "founder", "mcuban"
    "fly","initially","aside","press","center","tap","student","finally","schedule","home","owner","formally","invite","credit","arrive","explain","choose", "though","prepare","close"
    "elonmusk", "currently","provide","hire","inform","drive","publish","pepper","quit", "nadella","share","manage","scion", "graduate","order","tie","brush",
    "prefer","maybe","fail","paint","require", "estimate","together", "tout", "purchase", "closely", "since", "quote", "bachachan", 
     "clearly","hail","state","sama", "actually", "texte", "host", "may","product", "deal", "express","emerge", "investments",
     #---------------------------
    "allude", "strike", "grant", "throw", "love", "hear", "act", "apply", "regard", "maintain", "spot", "repeat", "court", "race", 
    "gather", "cause", "open", "wait", "rely", "rise", "catch", "delay", "round", "approach", "contact", "float", "spin", 
    "much", "regularly", "enough", "reckon", "afterward", "company", "designer", 
    "gates", "bezos", "zuckerberg", "page", "musk", "dorsey", "khosla", "pichai","alphabet","april","zaslav","fink","link", "february",
    #--------------------------
    "company", "native", "parent", "designer", "gates", "bezos", "zuckerberg","page", "musk", "dorsey", "khosla", "timcook",
     "hear","prove", "complain", "regard", "maintain", "spot", "repeat", "court", "race",
    "gather", "cause", "open", "wait", "rely", "contact", "float", "invoke", "shift", "spin", "pass",
    "throng", "act", "mean", "cook", "live" 
}
print(len(custom_stopwords))

nltk_stopwords = set(stopwords.words('english'))
full_stopwords = set(string.punctuation).union(nltk_stopwords)
full_stopwords = full_stopwords.union(custom_stopwords)
print(len(full_stopwords))
